+++ lib/shared/pages/analytics_page.dart
@@ -1,4 +1,5 @@
 import 'dart:async';
+import 'dart:convert';
 
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
@@ -9,10 +10,17 @@ import 'package:silversole/core/ble/ble_uuids.dart';
 import 'package:silversole/core/error/error_logger.dart';
 import 'package:silversole/core/error/result.dart';
 import 'package:silversole/core/utils/useful_extension.dart';
+import 'package:silversole/shared/dialogs/basic_dialog.dart';
+import 'package:silversole/shared/models/record_imu_notify_data_model.dart';
+import 'package:silversole/shared/providers/auth_provider.dart';
 import 'package:silversole/shared/providers/ble_connection_provider.dart';
 import 'package:silversole/shared/providers/settings_provider.dart';
 import 'package:silversole/shared/providers/telemetry_process_providers/live_telemetry_notifier.dart';
-import 'package:silversole/shared/widgets/chard_section.dart';
+import 'package:silversole/shared/widgets/chart/chart_section.dart';
+import 'package:silversole/shared/widgets/chart/imu_chart_section.dart';
+import 'package:silversole/shared/widgets/chart/record_imu_chart_section.dart';
+import 'package:silversole/shared/widgets/create_file_and_share.dart';
+import 'package:silversole/shared/widgets/update_check_bottom_modal.dart';
 
 class AnalyticsPage extends ConsumerStatefulWidget {
   const AnalyticsPage({super.key});
@@ -100,7 +108,7 @@ class _AnalyticsPageState extends ConsumerState<AnalyticsPage> {
       case Error():
         showErrorSnakeBar(result.error.toString());
       case Ok():
-        showMessage('start_record'.tr());
+        break;
     }
 
     unawaited(() async {
@@ -131,7 +139,6 @@ class _AnalyticsPageState extends ConsumerState<AnalyticsPage> {
       case Error():
         showErrorSnakeBar(result.error.toString());
       case Ok():
-        showMessage('stop_record'.tr());
         setState(() => _isRecording = false);
     }
 
@@ -162,14 +169,87 @@ class _AnalyticsPageState extends ConsumerState<AnalyticsPage> {
     }
   }
 
-  Future<void> exportRecorded() async {
-    comingSoon(); //TODO: complete export recorded json file
+  /// [example]
+  /// The output json string like:
+  /// ```
+  /// {
+  ///   "start_ts_ms": 1700000000000,
+  ///   "sample_rate_hz": 50,
+  ///   "samples": [
+  ///     [ts_ms, ax, ay, az, gx, gy, gz, pressure, wear_status, battery_percent],
+  ///     [...],
+  ///     ...
+  ///   ]
+  /// }
+  /// ```
+  String buildRecordJson(List<RecordImuNotifyDataModel> recordList) {
+    final startTs = recordList.first.timestamp;
+    final sampleRate = 200;
+    final samples = recordList
+        .map((e) => [e.timestamp, e.ax, e.ay, e.az, e.gx, e.gy, e.gz, e.pressure, e.wearStatus, e.batteryPercent])
+        .toList(growable: false);
+
+    return jsonEncode({
+      'start_ts_ms': startTs,
+    'sample_rate_hz': sampleRate,
+    'samples': samples
+    });
+  }
+
+  Future<void> exportRecorded(List<RecordImuNotifyDataModel> recordList) async {
+    if (recordList.isEmpty) return;
+    final userID = ref.read(authUserProvider)?.uuid ?? '';
+    final now = DateTime.now().toIso8601String().replaceAll(':', '-');
+    final json = buildRecordJson(recordList);
+    final optionsMap = {
+      'walk': 'walk'.tr(),
+      'run': 'run'.tr(),
+      'sitDown': 'sit_down'.tr(),
+      'standUp': 'stand_up'.tr(),
+      'fall': 'fall'.tr(),
+      'kick': 'kick'.tr()
+    };
+
+    await showContentDialog(
+      context,
+      title: 'export_record_file_title'.tr(),
+      content: ConstrainedBox(
+        constraints: BoxConstraints(
+          maxHeight: MediaQuery.sizeOf(context).height * 0.6,
+        ),
+        child: SingleChildScrollView(
+          child: Column(
+            spacing: 8,
+            mainAxisSize: MainAxisSize.min,
+            children: [
+              ...optionsMap.entries.map((e) {
+                final typeLabel = e.key;
+                final title = e.value;
+                return outlineButtonWithTheme(
+                  title: title,
+                  icon: LucideIcons.personStanding,
+                  onPressed: () {
+                    Navigator.pop(context);
+                    createFileAndShare(fileName: '${userID}_${now}_$typeLabel.json', content: json);
+                  },
+                );
+              }),
+            ],
+          ),
+        ),
+      )
+    );
+  }
+
+  Future<void> openShareFileBottomSheet() async {
+
   }
 
   @override
   Widget build(BuildContext context) {
-    final recordCount = ref.watch(liveTelemetryProvider).record.length;
-    final labels = ['pressure'.tr(), 'acc'.tr(), 'gyro'.tr(), 'battery'.tr()];
+    final recordList = ref.watch(liveTelemetryProvider).record;
+    final recordCount = recordList.length;
+    final labels = ['pressure'.tr(), 'acc'.tr(), 'gyro'.tr(), 'six-axis'.tr(), 'battery'.tr()];
 
     return Scaffold(
       appBar: AppBar(
@@ -189,8 +269,7 @@ class _AnalyticsPageState extends ConsumerState<AnalyticsPage> {
                   selectedIndex: _selectedIndex,
                   onSelected: (i) => setState(() => _selectedIndex = i),
                 ),
-                ChardSection(type: ChardDisplayType.single, selectedList: [_selectedIndex]),
-                const SizedBox(height: 16),
+                ImuChartSection(type: ChardDisplayType.single, selectedList: [_selectedIndex]),
                 Row(
                   spacing: 12,
                   children: [
@@ -233,20 +312,22 @@ class _AnalyticsPageState extends ConsumerState<AnalyticsPage> {
                         spacing: 32,
                         crossAxisAlignment: CrossAxisAlignment.start,
                         children: [
-                          Text('recording'.tr(), style: context.tt.titleSmall.bold),
-                          if (_isRecording)
-                          ChardSection(type: ChardDisplayType.all, source: ChardDataSource.recordImu),
+                          Text('record'.tr(), style: context.tt.titleSmall.bold),
+                          if (_isRecording) RecordImuChartSection(type: ChardDisplayType.all),
                           // if (_isRecording)
                           //   Text('already_recording_with_seconds'.tr(args: ["0"]), style: context.tt.titleMedium),
-                          Text('already_recording_with_count'.tr(args: [recordCount.toString()]), style: context.tt.titleMedium),
+                          Text(
+                            'already_recording_with_count'.tr(args: [recordCount.toString()]),
+                            style: context.tt.titleMedium,
+                          ),
                           if (!_isRecording)
                             Row(
                               mainAxisAlignment: MainAxisAlignment.end,
                               children: [
                                 FilledButton(
                                   style: FilledButton.styleFrom(backgroundColor: context.cs.primary, elevation: 0),
-                                  onPressed: exportRecorded,
-                                  child: Text('export'.tr()),
+                                  onPressed: recordList.isEmpty ? null : () => exportRecorded(recordList),
+                                  child: Row(spacing: 8, children: [Icon(LucideIcons.download), Text('export'.tr())]),
                                 ),
                               ],
                             ),